Add delete functionality for each list.

+++ app.js
@@ -105,15 +105,27 @@ app.post("/", (req, res) => {
 
 app.post("/delete", (req, res) => {
   const checkedItemId = req.body.checkbox;
+  const listName = req.body.listName;
 
-  Item.findByIdAndRemove(checkedItemId, (err) => {
-    if (err) {
-      console.log(err);
-    } else {
-      console.log("Deleted checked item");
-      res.redirect("/");
-    }
-  });
+  if (listName === "Today") {
+    Item.findByIdAndRemove(checkedItemId, (err) => {
+      if (err) {
+        console.log(err);
+      } else {
+        console.log("Deleted checked item");
+        res.redirect("/");
+      }
+    });
+  } else {
+    List.findOneAndUpdate(
+      {name: listName}, 
+      {$pull: {items: {_id: checkedItemId}}}, 
+      (err, result) => {
+        if (!err) {
+          res.redirect("/" + listName);
+        }
+      });
+  }
 });
 
 app.get("/about", (req, res) => {
